loan calc passed

+++ 20190926/activities/04-Loan-Calculator/starter-code/LoanCalculator/src/main/java/com/company/LoanCalculator.java
@@ -7,26 +7,26 @@ public class LoanCalculator {
         Scanner scanner = new Scanner(System.in);
 
         System.out.println("Enter Amount of Mortgage: ");
-        float L = scanner.nextFloat();
+        double L = scanner.nextDouble();
 
         System.out.println("Enter Term of Mortgage: ");
         int n = scanner.nextInt();
 
         System.out.println("Enter Annual Interest Rate of Mortgage");
-        float interest = scanner.nextFloat() / 100;
+        double interest = scanner.nextFloat() / 100d;
 
-        float c =  interest / 12.0000f;
+        double c =  interest / 12.0000d;
 
         // P = L[c(1 + c)^n]/[(1 + c)^n - 1]
         // P = Fixed Monthly Payment, L = Total Dollars, n = number of months, c = monthly interest rate
-        float onePlusC = 1.00f + c;
+        double onePlusC = 1.00d + c;
 
-        float numeratorExp = (float) Math.pow(onePlusC, n);
-        float numeratorExpTimesC = numeratorExp * c;
-        float numerator = L * numeratorExpTimesC;
-        float denominatorExp = (float) Math.pow((onePlusC), n);
-        float denominator = denominatorExp - 1;
-        float total = numerator / denominator;
+        double numeratorExp = (double) Math.pow(onePlusC, n);
+        double numeratorExpTimesC = numeratorExp * c;
+        double numerator = L * numeratorExpTimesC;
+        double denominatorExp = (double) Math.pow((onePlusC), n);
+        double denominator = denominatorExp - 1;
+        double total = numerator / denominator;
 
         System.out.println("Calculated Monthly Payment: " + total);
     }
